pyproject.toml
[tool.poetry]
name = "fooddelivery"
version = "0.1.0"
description = ""
authors = ["daniiar <daniiarz.test@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
pillow = "^9.5.0"
whitenoise = "^6.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
